+++ server-core/src/main/scala/com/convergencelabs/server/domain/model/OpenModelMetaData.scala
@@ -5,6 +5,6 @@ import java.time.Instant
 case class OpenModelMetaData(
   id: String,
   collection: String,
-  version: Int,
+  version: Long,
   createdTime: Instant,
   modifiedTime: Instant)
